requirements.txt
cryptography==1.2.3
requests==2.18.4
numpy==1.13.3
mxnet==1.1.0
turicreate==4.3.2
opencv-python==3.4.1.15
protobuf==3.5.2.post1
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.3
pymysql==0.9.2
